backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==10.1.0
PyMuPDF==1.23.8
pypdf==3.17.4
pdfplumber==0.10.3
spacy==3.7.2
openai==1.3.7
img2pdf==0.4.4
reportlab==4.0.7
opencv-python==4.8.1.78
numpy==1.24.3
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
